moving commands to utils file to keep main.go lighter.
