+++ cmd/wallet/main.go
@@ -2,15 +2,12 @@ package main
 
 import (
 	"bufio"
-	"encoding/json"
 	"fmt"
 	"log"
 	"os"
-	"strconv"
 	"strings"
 
 	"github.com/Maphikza/btc-wallet-btcsuite.git/internal/config"
-	"github.com/Maphikza/btc-wallet-btcsuite.git/internal/ipc"
 	"github.com/Maphikza/btc-wallet-btcsuite.git/internal/logger"
 	setupwallet "github.com/Maphikza/btc-wallet-btcsuite.git/internal/wallet"
 	_ "github.com/btcsuite/btcwallet/walletdb/bdb"
@@ -116,296 +113,4 @@ func interactiveMode() {
 			fmt.Println("Invalid choice. Please try again.")
 		}
 	}
-}
-
-var createWalletCmd = &cobra.Command{
-	Use:   "create [wallet-name] [password] [pubKey] [apiKey]",
-	Short: "Create a new wallet",
-	Long: `Create a new wallet with the given name and password. 
-	Optionally provide a pubKey and apiKey for panel integration.`,
-	Args: cobra.RangeArgs(2, 4),
-	Run: func(cmd *cobra.Command, args []string) {
-		walletName := args[0]
-		password := args[1]
-		pubKey := ""
-		apiKey := ""
-		if len(args) > 2 {
-			pubKey = args[2]
-		}
-		if len(args) > 3 {
-			apiKey = args[3]
-		}
-
-		mnemonic, err := setupwallet.CreateWalletAPI(walletName, password, pubKey, apiKey)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error creating wallet: %v\n", err)
-			os.Exit(1)
-		}
-
-		result := struct {
-			WalletName string `json:"walletName"`
-			Mnemonic   string `json:"mnemonic"`
-		}{
-			WalletName: walletName,
-			Mnemonic:   mnemonic,
-		}
-
-		json.NewEncoder(os.Stdout).Encode(result)
-	},
-}
-
-var importWalletCmd = &cobra.Command{
-	Use:   "import [wallet-name] [mnemonic] [password] [birthdate] [pubKey] [apiKey]",
-	Short: "Import an existing wallet",
-	Long: `Import an existing wallet with the given name, mnemonic, and password. 
-	Provide the wallet's birthdate in YYYY-MM-DD format.
-	Optionally provide a pubKey and apiKey for panel integration.`,
-	Args: cobra.RangeArgs(4, 6),
-	Run: func(cmd *cobra.Command, args []string) {
-		walletName := args[0]
-		mnemonic := args[1]
-		password := args[2]
-		birthdate := args[3]
-		pubKey := ""
-		apiKey := ""
-		if len(args) > 4 {
-			pubKey = args[4]
-		}
-		if len(args) > 5 {
-			apiKey = args[5]
-		}
-
-		err := setupwallet.ImportWalletAPI(walletName, mnemonic, password, birthdate, pubKey, apiKey)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error importing wallet: %v\n", err)
-			os.Exit(1)
-		}
-
-		result := struct {
-			WalletName string `json:"walletName"`
-			Message    string `json:"message"`
-		}{
-			WalletName: walletName,
-			Message:    "Wallet imported successfully",
-		}
-
-		json.NewEncoder(os.Stdout).Encode(result)
-	},
-}
-
-var openWalletCmd = &cobra.Command{
-	Use:   "open [wallet-name] [password]",
-	Short: "Open and load a wallet",
-	Long:  `Open and load a wallet with the given name and password.`,
-	Args:  cobra.ExactArgs(2),
-	Run: func(cmd *cobra.Command, args []string) {
-		walletName := args[0]
-		password := args[1]
-
-		baseDir, err := os.Getwd()
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error getting current directory: %v\n", err)
-			os.Exit(1)
-		}
-
-		err = setupwallet.OpenAndLoadWalletAPI(walletName, password, baseDir)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error opening wallet: %v\n", err)
-			os.Exit(1)
-		}
-
-		result := struct {
-			WalletName string `json:"walletName"`
-			Message    string `json:"message"`
-		}{
-			WalletName: walletName,
-			Message:    "Wallet opened successfully",
-		}
-
-		json.NewEncoder(os.Stdout).Encode(result)
-	},
-}
-
-var newTransactionCmd = &cobra.Command{
-	Use:   "new-transaction [recipient] [amount] [fee-rate]",
-	Short: "Create a new transaction",
-	Long:  `Create a new transaction with the specified recipient, amount (in satoshis), and fee rate (in sat/vB).`,
-	Args:  cobra.ExactArgs(3),
-	Run: func(cmd *cobra.Command, args []string) {
-		client, err := ipc.NewClient()
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error connecting to wallet server: %v\n", err)
-			os.Exit(1)
-		}
-		defer client.Close()
-
-		result, err := client.SendCommand("new-transaction", args)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error creating transaction: %v\n", err)
-			os.Exit(1)
-		}
-
-		json.NewEncoder(os.Stdout).Encode(result)
-	},
-}
-
-var rbfTransactionCmd = &cobra.Command{
-	Use:   "rbf-transaction [original-txid] [new-fee-rate]",
-	Short: "Replace a transaction with a higher fee",
-	Long:  `Replace an existing transaction with a new one that has a higher fee rate (in sat/vB).`,
-	Args:  cobra.ExactArgs(2),
-	Run: func(cmd *cobra.Command, args []string) {
-		client, err := ipc.NewClient()
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error connecting to wallet server: %v\n", err)
-			os.Exit(1)
-		}
-		defer client.Close()
-
-		result, err := client.SendCommand("rbf-transaction", args)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error performing RBF transaction: %v\n", err)
-			os.Exit(1)
-		}
-
-		log.Println("RBF result: ", result)
-
-		json.NewEncoder(os.Stdout).Encode(result)
-	},
-}
-
-var getWalletBalanceCmd = &cobra.Command{
-	Use:   "balance",
-	Short: "Get the current wallet balance",
-	Long:  `Retrieve the current balance of the opened wallet.`,
-	Args:  cobra.NoArgs,
-	Run: func(cmd *cobra.Command, args []string) {
-		client, err := ipc.NewClient()
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error connecting to wallet server: %v\n", err)
-			os.Exit(1)
-		}
-		defer client.Close()
-
-		result, err := client.SendCommand("get-wallet-balance", nil)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error getting wallet balance: %v\n", err)
-			os.Exit(1)
-		}
-		log.Println("Wallet Balance results: ", result)
-
-		json.NewEncoder(os.Stdout).Encode(result)
-	},
-}
-
-var estimateTransactionSizeCmd = &cobra.Command{
-	Use:   "estimate-tx-size [spend-amount] [recipient-address] [fee-rate]",
-	Short: "Estimate transaction size",
-	Long:  `Estimate the size of a transaction with the given parameters.`,
-	Args:  cobra.ExactArgs(3),
-	Run: func(cmd *cobra.Command, args []string) {
-		client, err := ipc.NewClient()
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error connecting to wallet server: %v\n", err)
-			os.Exit(1)
-		}
-		defer client.Close()
-
-		spendAmount, err := strconv.ParseInt(args[0], 10, 64)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Invalid spend amount: %v\n", err)
-			os.Exit(1)
-		}
-
-		feeRate, err := strconv.Atoi(args[2])
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Invalid fee rate: %v\n", err)
-			os.Exit(1)
-		}
-
-		result, err := client.SendCommand("estimate-transaction-size", []string{
-			strconv.FormatInt(spendAmount, 10),
-			args[1], // recipient address
-			strconv.Itoa(feeRate),
-		})
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error estimating transaction size: %v\n", err)
-			os.Exit(1)
-		}
-
-		json.NewEncoder(os.Stdout).Encode(result)
-	},
-}
-
-var getTransactionHistoryCmd = &cobra.Command{
-	Use:   "tx-history",
-	Short: "Get transaction history",
-	Long:  `Retrieve the transaction history of the opened wallet.`,
-	Args:  cobra.NoArgs,
-	Run: func(cmd *cobra.Command, args []string) {
-		client, err := ipc.NewClient()
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error connecting to wallet server: %v\n", err)
-			os.Exit(1)
-		}
-		defer client.Close()
-
-		result, err := client.SendCommand("get-transaction-history", nil)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error getting transaction history: %v\n", err)
-			os.Exit(1)
-		}
-
-		log.Println("Transaction History: ", result)
-
-		json.NewEncoder(os.Stdout).Encode(result)
-	},
-}
-
-var getReceiveAddressesCmd = &cobra.Command{
-	Use:   "get-receive-addresses",
-	Short: "Get all generated receive addresses from the wallet",
-	Long:  "Retrieve a list of all generated receive addresses from the wallet.",
-	Args:  cobra.NoArgs,
-	Run: func(cmd *cobra.Command, args []string) {
-		client, err := ipc.NewClient()
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error connecting to wallet server: %v\n", err)
-			os.Exit(1)
-		}
-		defer client.Close()
-
-		result, err := client.SendCommand("get-receive-addresses", nil)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error getting receive addresses: %v\n", err)
-			os.Exit(1)
-		}
-
-		json.NewEncoder(os.Stdout).Encode(result)
-	},
-}
-
-var exitWalletCmd = &cobra.Command{
-	Use:   "exit",
-	Short: "Exit and shut down the wallet",
-	Long:  "Gracefully shut down the wallet.",
-	Args:  cobra.NoArgs,
-	Run: func(cmd *cobra.Command, args []string) {
-		client, err := ipc.NewClient()
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error connecting to wallet server: %v\n", err)
-			os.Exit(1)
-		}
-		defer client.Close()
-
-		result, err := client.SendCommand("exit", nil)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error exiting wallet: %v\n", err)
-			os.Exit(1)
-		}
-
-		log.Println("Exit Result: ", result)
-
-		json.NewEncoder(os.Stdout).Encode(result)
-	},
 }
